Add input_type support on Android using SDL2
The "show_keyboard" method located in the android module stopped working
when SDL2 was implemented as a backend in Kivy. This because the SDL2
implementation did not include a way to handle the "input_type"
property, which can be defined by the user in the "TextInput" class.

As it depends directly on the Kivy repository, some changes were made
there to let the "_WindowSDL2Storage.show_keyboard" method accept the
"input_type" value and call the "android.show_keyboard" function.

Nonetheless, the android.show_keyboard did not handle properly the
different input types that can be set in the TextInput class (text,
number, url, mail, datetime, tel and address), including the password
variation of text and number, and the keyboard suggestions for text,
url, mail and address. This commit solves that by adding the missing
flags, removing the unused and adding the missing logic.

The reason the if statement related to the "keyboard_suggestions" is in
the same block than the one checking if "password" is True is because a
keyboard showing suggestions in a password field is not compatible, so
one of them can be set, but never both.